UGNI: split files into new sub directories, create common iface class
