+++ src/uct/ugni/base/ugni_pd.c
@@ -20,19 +20,6 @@ static ucs_status_t uct_ugni_pd_open(const char *pd_name, uct_pd_h *pd_p);
 UCS_CONFIG_DEFINE_ARRAY(ugni_alloc_methods, sizeof(uct_alloc_method_t),
                         UCS_CONFIG_TYPE_ENUM(uct_alloc_method_names));
 
-ucs_config_field_t uct_ugni_iface_config_table[] = {
-    /* This tuning controls the allocation priorities for bouncing buffers */
-    { "", "MAX_SHORT=2048;MAX_BCOPY=2048;ALLOC=huge,mmap,heap", NULL,
-    ucs_offsetof(uct_ugni_iface_config_t, super), UCS_CONFIG_TYPE_TABLE(uct_iface_config_table)},
-
-    UCT_IFACE_MPOOL_CONFIG_FIELDS("FMA", -1, 0, "fma",
-                                  ucs_offsetof(uct_ugni_iface_config_t, mpool),
-                                  "\nAttention: Setting this param with value != -1 is a dangerous thing\n"
-                                  "and could cause deadlock or performance degradation."),
-
-    {NULL}
-};
-
 pthread_mutex_t uct_ugni_global_lock = PTHREAD_MUTEX_INITIALIZER;
 
 /* For Cray devices we have only one PD */
